add featur comment and favorites

+++ backend/src/controll/comment/comment.services.js
@@ -1,5 +1,5 @@
 // src/controll/comment/comment.services.js
-const { findCommentsByWisataId, createComment, deleteComment } = require("./comment.repository.js");
+const { findCommentsByWisataId, createComment, deleteComment, findCommentById, updateComment } = require("./comment.repository.js");
 
 const getCommentsByWisataId = async (wisataId) => {
   return await findCommentsByWisataId(wisataId);
@@ -9,12 +9,33 @@ const addComment = async (commentData) => {
   return await createComment(commentData);
 };
 
-const removeComment = async (id) => {
-  return await deleteComment(id);
+const editComment = async (id, commentData, userId, userRole) => {
+  const comment = await findCommentById(id);
+
+  if (!comment) {
+    throw new Error('Comment not found');
+  }
+
+  if (comment.userId !== userId && userRole !== 'ADMIN') {
+    throw new Error('Unauthorized');
+  }
+
+  return await updateComment(id, commentData);
+};
+
+const removeComment = async (id, userId, userRole) => {
+  const comment = await findCommentById(id);
+
+  if (comment.userId !== userId && userRole !== "ADMIN") {
+    throw new Error("Unauthorized");
+  }
+
+  await deleteComment(id);
 };
 
 module.exports = {
   getCommentsByWisataId,
   addComment,
   removeComment,
+  editComment
 };
